# src/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Yet Another Node.js Blog Application API",
    "description": "Yet Another Node.js Blog Application API",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "host": "localhost:3000",
  "basePath": "/api/v1",
...
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/restaurant": {
      "get": {
        "tags": [
          "Restaurant"
        ],
        "description": "List of all restaurants",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "page number, Default to Zero",
            "type": "integer"
          },
          {
            "name": "order",
            "in": "query",
            "description": "DESC or ASC for rating, default to DSC",
            "type": "integer"
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "List of restaurants with page size of 50",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string"
                },
                "data": {
                  "type": "array",
                  "items": {
                    "$ref": "#/definitions/Restaurant"
                  }
                }
              }
            }
          },
          "4xx": {
            "description": "Fetching failed",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      },
      "post": {
        "tags": [
          "Restaurant"
        ],
        "description": "Create new restaurant in system",
        "parameters": [
          {
            "name": "restaurant",
            "in": "body",
            "description": "Restaurant that we want to create",
            "schema": {
              "properties": {
                "rating": {
                  "type": "string",
                  "example": 4.5
                },
                "address": {
                  "type": "string",
                  "example": "address"
                },
                "name": {
                  "type": "string",
                  "example": "restaurant name"
                }
              }
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "201": {
            "description": "New restaurant is created",
            "schema": {
              "$ref": "#/definitions/SuccessUserResponse"
            }
          },
          "40x": {
            "description": "When restaurant creation is failed",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },
    "/restaurant/search": {
      "parameters": [
        {
          "name": "rating",
          "in": "body",
          "description": "rating on the basis of which you want to search restaurants",
          "schema": {
            "properties": {
              "name": {
                "type": "string",
                "example": "name"
              },
              "ids": {
                "type": "array",
                "example": "[1,2,3,4,5]"
              },
              "rating": {
                "type": "array",
                "example": "[3.4, 4.5, 4.5]"
              }
            }
          }
        }
      ],
      "post": {
        "tags": [
          "Restaurant"
        ],
        "summary": "Get restaurants with given ID",
        "responses": {
          "200": {
            "description": "Restaurants are found",
            "schema": {
              "properties": {
                "success": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/definitions/Restaurant"
                      }
                    }
                  }
                }
              }
            }
          },
          "4xx": {
            "description": "Restaurant is not found",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },
    "/restaurant/{restaurantId}": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "description": "ID of restaurant",
          "type": "string"
        }
      ],
      "get": {
        "tags": [
          "Restaurant"
        ],
        "summary": "Get restaurant with given restaurant ID",
        "responses": {
          "200": {
            "description": "Restaurant is found",
            "schema": {
              "properties": {
                "success": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "$ref": "#/definitions/Restaurant"
                    }
                  }
                }
              }
            }
          },
          "4xx": {
            "description": "Restaurant is not found",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      },
      "put": {
        "summary": "Update restaurant with a given ID",
        "tags": [
          "Restaurant"
        ],
        "parameters": [
          {
            "name": "restaurant",
            "in": "body",
            "description": "values that you want to update",
            "schema": {
              "properties": {
                "name": {
                  "type": "string",
                  "example": "name"
                },
                "address": {
                  "type": "string",
                  "example": "address"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Restaurant is Updated",
            "schema": {
              "properties": {
                "success": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "$ref": "#/definitions/Restaurant"
                    }
                  }
                }
              }
            }
          },
          "4xx": {
            "description": "error in updating restaurant",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Restaurant"
        ],
        "summary": "Delete restaurant with given ID ( only admin and manager can )",
        "responses": {
          "200": {
            "description": "Restaurant is Deleted",
            "schema": {
              "properties": {
                "success": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "$ref": "#/definitions/Restaurant"
                    }
                  }
                }
              }
            }
          },
          "4xx": {
            "description": "error in deleting restaurant",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "Restaurant": {
      "properties": {
        "id": {
          "type": "integer"
        },
        "rating": {
          "type": "string",
          "example": 4.5
        },
        "address": {
          "type": "string",
          "example": "address"
        },
        "name": {
          "type": "string",
          "example": "restaurant name"
        },
        "createdAt": {
          "type": "string",
          "example": "ISO Format date"
        },
        "updatedAt": {
          "type": "string",
          "example": "ISO Format date"
        }
      }
    },
...
    "SuccessUserResponse": {
      "type": "object",
      "properties": {
        "success": {
          "type": "object",
          "properties": {
            "message": {
              "type": "string",
              "description": "Success message"
            }
          }
        }
      }
    },
    "ErrorResponse": {
      "type": "object",
      "properties": {
        "error": {
          "type": "object",
          "properties": {
            "message": {
              "type": "string"
            }
          }
        }
      }
    }
  }
}